extract read rows to own package
