+++ agent/runner/actions/postgresql_show_index_action.go
@@ -27,6 +27,7 @@ import (
 
 	"github.com/percona/pmm/agent/utils/templates"
 	"github.com/percona/pmm/api/agentpb"
+	"github.com/percona/pmm/utils/sqlrows"
 )
 
 type postgresqlShowIndexAction struct {
@@ -92,7 +93,7 @@ func (a *postgresqlShowIndexAction) Run(ctx context.Context) ([]byte, error) {
 		return nil, errors.WithStack(err)
 	}
 
-	columns, dataRows, err := readRows(rows)
+	columns, dataRows, err := sqlrows.ReadRows(rows)
 	if err != nil {
 		return nil, errors.WithStack(err)
 	}
